플레이스 홀더 치환 및 데이터 변환 TransformationHelper개선

+++ src/main/java/com/example/cmtProject/mapper/erp/eapproval/DocumentMapper.java
@@ -116,7 +116,7 @@ public interface DocumentMapper {
     /**
      * 직위 번호로 직위명 조회 
      */
-	public String selectPositionNameByPositionNo(Integer positionNo);
+	public String selectPositionNameByPositionNo(Long positionNo);
 	
 	
 }
